test(ai): isolate integration seams
