+++ packages/pyric-tools/test/serve/vite-plugin.test.ts
@@ -238,81 +238,6 @@ describe('integration — real vite dev pluginContainer', () => {
   });
 });
 
-describe('integration — canonical AI operation through Vite', () => {
-  it('shares the default app handle and completes generateContent()', async () => {
-    const root = mkdtempSync(join(tmpdir(), 'pyric-vite-ai-'));
-    writeFileSync(
-      join(root, 'ai-smoke.ts'),
-      `import { initializeApp } from 'firebase/app';
-import { getAI, getGenerativeModel } from 'firebase/ai';
-
-export async function run() {
-  const app = initializeApp({ apiKey: 'ignored', projectId: 'vite-ai-smoke' });
-  const explicit = getAI(app);
-  const bare = getAI();
-  const result = await getGenerativeModel(explicit, {
-    model: 'gemini-flash-lite-latest',
-  }).generateContent('hello');
-  return {
-    sameHandle: explicit === bare,
-    location: explicit.location,
-    candidates: result.response.candidates.length,
-  };
-}
-`,
-    );
-
-    const originalFetch = globalThis.fetch;
-    const runtimeGlobal = globalThis as typeof globalThis & {
-      __PYRIC_FORCE_INPAGE__?: boolean;
-    };
-    const originalForceInPage = runtimeGlobal.__PYRIC_FORCE_INPAGE__;
-    runtimeGlobal.__PYRIC_FORCE_INPAGE__ = true;
-    globalThis.fetch = (async (input: RequestInfo | URL, init?: RequestInit) => {
-      if (String(input) === '/__pyric/init.json') {
-        return Response.json({
-          rules: null,
-          rulesHash: null,
-          bridgeUrl: null,
-          seed: null,
-          persist: false,
-        });
-      }
-      return originalFetch(input, init);
-    }) as typeof fetch;
-
-    const { createServer } = await import('vite');
-    let server: Awaited<ReturnType<typeof createServer>> | undefined;
-    try {
-      server = await createServer({
-        appType: 'custom',
-        configFile: false,
-        logLevel: 'silent',
-        root,
-        plugins: [pyricSandbox({ ui: false })],
-        server: { middlewareMode: true },
-        optimizeDeps: { noDiscovery: true },
-      });
-      const fixture = await server.ssrLoadModule('/ai-smoke.ts') as {
-        run(): Promise<{ sameHandle: boolean; location: string; candidates: number }>;
-      };
-      await expect(fixture.run()).resolves.toEqual({
-        sameHandle: true,
-        location: '',
-        candidates: 1,
-      });
-    } finally {
-      await server?.close();
-      globalThis.fetch = originalFetch;
-      if (originalForceInPage === undefined) {
-        delete runtimeGlobal.__PYRIC_FORCE_INPAGE__;
-      } else {
-        runtimeGlobal.__PYRIC_FORCE_INPAGE__ = originalForceInPage;
-      }
-    }
-  }, 60_000);
-});
-
 // ── integration: the BRIDGE mounts inside a REAL Vite dev server (middlewareMode,
 // binds NO port). The unit tests above drive a hand-rolled stub; this proves the
 // plugin actually registers its `/__pyric` middleware into real Vite's connect
